libft refatorada

+++ ft_strncmp.c
@@ -6,7 +6,7 @@
 /*   By: bda-mota <bda-mota@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/11 19:48:18 by bda-mota          #+#    #+#             */
-/*   Updated: 2023/10/16 19:32:09 by bda-mota         ###   ########.fr       */
+/*   Updated: 2023/10/25 20:45:46 by bda-mota         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,5 +21,5 @@ int	ft_strncmp(const char *s1, const char *s2, size_t n)
 		return (0);
 	while (i < (n - 1) && s1[i] && s2[i] && s1[i] == s2[i])
 		i++;
-	return (((int)(unsigned char)s1[i]) - ((int)(unsigned char)s2[i]));
+	return ((unsigned char)s1[i] - (unsigned char)s2[i]);
 }
